+++ src/main/java/com/dsmp/service/SubjectService.java
@@ -5,5 +5,5 @@ public interface SubjectService {
      * @param subId 科目id
      * @return
      */
-	public Integer findNeedStudyTime(Integer subId);
+	public Double findNeedStudyTime(Integer subId);
 }
