+++ src/main/java/com/dsmp/service/impl/SubjectServiceImpl.java
@@ -14,8 +14,8 @@ public class SubjectServiceImpl implements SubjectService {
      * @return
      */
 	@Override
-	public Integer findNeedStudyTime(Integer subId) {
-		Integer needStyTime = subjectMapper.findNeedStudyTime(subId);
+	public Double findNeedStudyTime(Integer subId) {
+		Double needStyTime = subjectMapper.findNeedStudyTime(subId);
 //		System.out.println("needStyTime-service:"+needStyTime);
 //		 int i = (needStyTime*60*60).intValue();
 //		Integer needTime = Integer.parseInt(new java.text.DecimalFormat("0").format(needStyTime*60*60));
